Added: automatically moves .rar files to the top of the queue. Needs testing as it may mess with queues that contain a lot of .rar files.
Added: ignore_list Similar to cleanup_list but does not add any files to the queue that match what is in the cleanup_list. IE ".sample, -sample" removed *.sample* and *-sample* Needs to be checked if it removes too much such as if there are any posts that have "-samples" in.