Do explicit handling of login problems in downloader.py. Two cases are handled: - Failed login: server will be blocked until its parameters change - Too many connections: server thread count will be lowered until situation is correct (not permanent!)
Lack of standardization makes implementation tricky.
For login-failure, servers use 481 (official) or 502 and a server-specific message.
For too many connections, servers use 400 or 502 and a server-specific message.
The code tries to look for clues in the message combined with the response code.
When no conclusion can be drawn, the error is seen as temporary.

Closes #182 for trunk.