Improved duplicate detection.
The generic version is based on the md5sum of the combined article-identifiers and no longer on the NZB backup.
Episode detection is based on the  name/season/episode extraction from the job title.
There's now an option for generic detection and an option for episode detection.
All info is stored in the History database, under a new user_version number "1" because two new columns are needed.