Adding OZnzb to Glitter
Changed OZnzb so it will allow changing of up/down voting.
For out of retention it will only list the user's servers, not all of them.
In Glitter you can only give a audio/video score once (just like on the website).
Plush allows changing of audio/video rating and shows that change, but it is not actually changed on the website.