Adding central speedhistory
Adding a central speed history within SAB, a feature many download
programs have and I think SAB could use! This way you can actually see
what happened when the browser was not open.
When there is no download-activity nothing really happens, it only calls
and updates (in case 0's need to be added) when the main.tmpl file is
called. In Glitter this is only once.
It does add consecutive numbers to a list, but this should not be a
memory hog.