Integrating icon font Woff as Base64 into CSS for Config and Glitter
This fixes any problems that Chrome has with fonts on auth/https
connections that don't have valid certificates. Both Firefox and Chrome
use Woff font's, so only for IE it will have to download 20kb extra of
the CSS while for FF/Chrome it will lower the number of requests by 1.