Correctly handle obfuscated filenames
MultiPar will always rename obfuscated that are complete. So if the repair then fails because some are incomplete, it will leave those with the old filenames.
So on the next attempt (after fetching blocks) we need to make sure we also tell SAB about the old and the new changes!
Otherwise we get errors at Unpack about missing files, referring to the obfuscated ones.