Update encryption check to handle partially assembled files
https://forums.sabnzbd.org/viewtopic.php?f=3&p=123147
In SABnzbd 3.x we write incomplete files to the disk instead of waiting for a file to be 100% complete.
So the password check fails because it will check part001 and automatically continue to part002. Instead of crashing with a "can't find part002" (this we expect) it finds a incomplete part002 and crashes with a different error that we don't catch.
Alternatively it can crash due to starting to check on part002 while part001 isn't there yet. This used to work, but broke now.