Improve new Quick-check implementation
We only care about missing/broken articles in files that we have par2 for. So we check for each NZF if it has bad articles, and only fail if it is part of a par2 set. Additionally we check if the file size matches the one from par2.

Since this also enables CRC check for sabyenc3, it will be slower and the default of num_simd_decoders is increased to 2.