Refresh and expand the appstream metadata
Add supported control methods [1] as well as recommended screen sizes [2] and internet availability. Most of these properties are already in active use by appdata clients such a Gnome's "Software" program. The display size basically says "all but extra-small", where the extra extra-small category is used for tiny devices such a wearables and watches, and is kept as a recommend to not block installs on headless systems.

Also set a vcs-browser URL, and update the contact URL to point to the more generic live-chat page rather than directly to the forums.

[1] https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-relations-control
[2] https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-relations-display_length