Include byte unit in quota strings
Fixes #2590

I expect to continue to iterate on some of the ideas in the comments
thereon, but this fixes the proximate issue mentioned in the issue.

This includes changes to the interface to properly interpret the new
strings:

- Remove the UI-based `B` designations as they are now tagged with
  units from the api
- Let parseInt do its magic with string-y numbers plus units
- Expand parseInt for use in the checking whether quota is set

The display will be better in this `Glitter` case. This may well be
considered a breaking change for the API and not applied.